# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dynamic user segmentation service",
        "title": "Avito App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/priv": {
            "get": {
                "description": "Get All Privileges",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Privilege"
                ],
                "summary": "GetPrivilege",
                "operationId": "GetPrivilege",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PrivilegeDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PrivilegeResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "default": {
                        "description": ""
                    }
                }
            },
            "post": {
                "description": "Create Privileges",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Privilege"
                ],
                "summary": "CreatePrivilege",
                "operationId": "CreatePrivilege",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PrivilegeDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/priv/:id": {
            "delete": {
                "description": "Delete Privilege by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Privilege"
                ],
                "summary": "DeletePrivilege",
                "operationId": "DeletePrivilege",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/priv/user": {
            "get": {
                "description": "Get All Users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "GetAllUsers",
                "operationId": "GetUsers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PrivilegedUserDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/priv/user/:id": {
            "delete": {
                "description": "Delete User by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "DeleteUser",
                "operationId": "DeleteUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/priv/user/add": {
            "post": {
                "description": "Add Privileges to User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "AddPrivilegeToUser",
                "operationId": "AddPrivileges",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PrivilegedUserCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/priv/user/remove": {
            "post": {
                "description": "Remove Privileges to User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "RemovePrivilegeToUser",
                "operationId": "RemovePrivileges",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PrivilegedUserDeleteDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "default": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.PrivilegeDTO": {
            "type": "object",
            "required": [
                "privilege_title"
            ],
            "properties": {
                "privilege_title": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
        "dto.PrivilegeResponseDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "privilege_title": {
                    "type": "string"
                }
            }
        },
        "dto.PrivilegedUserCreateDTO": {
            "type": "object",
            "properties": {
                "add_privilege": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "dto.PrivilegedUserDTO": {
            "type": "object",
            "required": [
                "privilege_title"
            ],
            "properties": {
                "privilege_title": {
                    "type": "string",
                    "maxLength": 20
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "dto.PrivilegedUserDeleteDTO": {
            "type": "object",
            "properties": {
                "del_privilege": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}